services/wrappers-py/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
torch
transformers
sentencepiece
langchain-openai
langchain-core
